+++ tapis-systemslib/src/test/java/edu/utexas/tacc/tapis/systems/IntegrationUtils.java
@@ -248,10 +248,9 @@ public final class IntegrationUtils
   public static final boolean showDeletedTrue = true;
   public static final boolean resolveEffUserTrue = true;
   public static final boolean resolveEffUserFalse = false;
-  public static final boolean sharedAppCtxTrue = true;
-  public static final boolean sharedAppCtxFalse = false;
-  public static final String sharedAppCtxOwner = owner1;
-  public static final String sharedAppCtxNull = null;
+  public static final String sharedCtxOwner = owner1;
+  public static final String sharedCtxNull = null;
+  public static final String resourceTenantNull = null;
   public static final Instant createdNull = null;
   public static final Instant updatedNull = null;
   public static final int qMaxJobs = -1;